LICENSE
The peer production license is an example of the Copyfair type of license, in which only other commoners, cooperatives and nonprofits can share and re-use the material, but not commercial entities intent on making profit through the commons without explicit reciprocity. This fork on the original text of the Creative Commons non-commercial variant makes the PPL an explicitly anti-capitalist version of the CC-NC. It only allows commercial exploitation by collectives in which the ownership of the means of production is in the hands of the value creators, and where any surplus is distributed equally among them (and not only into the hands of owners, shareholders or absentee speculators). According to Dmytri Kleiner, co-author of the license with the barrister John Magyar, it’s not a copyleft license, but instead copyFARleft, and is intended for consumer goods or commodities rather than capital or producers’ goods.

Más información: https://wiki.p2pfoundation.net/Peer_Production_License